+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @alwaysmeticulous/cli
 
+## 2.334.0
+
+### Patch Changes
+
+- Updated dependencies [[`1163ef2`](https://github.com/alwaysmeticulous/meticulous/commit/1163ef2b5bfba9f62aaf65748f7e909d6a34ebda), [`a169be3`](https://github.com/alwaysmeticulous/meticulous/commit/a169be364e19cda80cd24b21181ab3c373ba59a0)]:
+  - @alwaysmeticulous/downloading-helpers@2.334.0
+  - @alwaysmeticulous/api@2.334.0
+  - @alwaysmeticulous/debug-workspace@2.334.0
+  - @alwaysmeticulous/replay-orchestrator-launcher@2.334.0
+  - @alwaysmeticulous/client@2.334.0
+  - @alwaysmeticulous/common@2.333.1
+  - @alwaysmeticulous/record@2.334.0
+  - @alwaysmeticulous/remote-replay-launcher@2.334.0
+  - @alwaysmeticulous/sdk-bundles-api@2.334.0
+  - @alwaysmeticulous/session-filters@2.334.0
+  - @alwaysmeticulous/replay-debugger-ui@2.333.1
+
 ## 2.333.2
 
 ### Patch Changes