+++ packages/session-filters/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @alwaysmeticulous/session-filters
 
+## 2.334.0
+
+### Patch Changes
+
+- Updated dependencies [[`a169be3`](https://github.com/alwaysmeticulous/meticulous/commit/a169be364e19cda80cd24b21181ab3c373ba59a0)]:
+  - @alwaysmeticulous/api@2.334.0
+
 ## 2.333.1
 
 ### Patch Changes